requirements.txt
flask
PyPDF2
scikit-learn
gunicorn
google-generativeai